backend/requirements.txt
flask>=3.0
flask-cors>=4.0
pyyaml>=6.0
watchdog>=4.0
